# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API for managing an online song library",
    "title": "Online Song Library",
    "contact": {},
    "version": "1.0"
  },
  "host": "localhost:8081",
  "basePath": "/",
  "paths": {
    "/api/songs": {
      "get": {
        "description": "Returns a list of songs with filtering and pagination",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "songs"
        ],
        "summary": "Get list of songs",
        "parameters": [
          {
            "type": "string",
            "description": "Group name",
            "name": "group",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Song name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Release date",
            "name": "releaseDate",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Song text",
            "name": "text",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Song link",
            "name": "link",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Offset",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          }
        }
      }
    },
    "/api/songs/new": {
      "post": {
        "description": "Adds a new song and saves it to the database",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "songs"
        ],
        "summary": "Add a new song",
        "parameters": [
          {
            "description": "New song",
            "name": "song",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.NewSongPayload"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/models.Song"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          }
        }
      }
    },
    "/api/songs/{song_id}/delete": {
      "delete": {
        "description": "Deletes a song by ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "songs"
        ],
        "summary": "Delete a song",
        "parameters": [
          {
            "type": "string",
            "description": "Song ID",
            "name": "song_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          }
        }
      }
    },
    "/api/songs/{song_id}/edit": {
      "patch": {
        "description": "Edits song data by ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "songs"
        ],
        "summary": "Edit song data",
        "parameters": [
          {
            "type": "string",
            "description": "Song ID",
            "name": "song_id",
            "in": "path",
            "required": true
          },
          {
            "description": "Song data",
            "name": "song",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.EditSongPayload"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          }
        }
      }
    },
    "/api/songs/{song_id}/text": {
      "get": {
        "description": "Returns the text of a song with pagination over verses",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "songs"
        ],
        "summary": "Get song text",
        "parameters": [
          {
            "type": "string",
            "description": "Song ID",
            "name": "song_id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "Page number",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Number of verses per page",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handlers.JSON"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "handlers.JSON": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "song": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.Song"
          }
        },
        "text": {
          "type": "string"
        }
      }
    },
    "models.EditSongPayload": {
      "type": "object",
      "properties": {
        "song": {
          "$ref": "#/definitions/models.Song"
        },
        "song_details": {
          "$ref": "#/definitions/models.SongDetails"
        }
      }
    },
    "models.NewSongPayload": {
      "type": "object",
      "properties": {
        "group": {
          "type": "string",
          "example": "Muse"
        },
        "song": {
          "type": "string",
          "example": "Supermassive Black Hole"
        }
      }
    },
    "models.Song": {
      "type": "object",
      "properties": {
        "group_id": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "models.SongDetails": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "link": {
          "type": "string"
        },
        "release_date": {
          "type": "string"
        },
        "song_id": {
          "type": "integer"
        },
        "text": {
          "type": "string"
        }
      }
    }
  }
}